chore: replace IdentityCheck icons with new format
