+++ src/features/cheatcodes/pages/AppComponents/AppComponents.tsx
@@ -88,6 +88,7 @@ import { Clock } from 'ui/svg/icons/Clock'
 import { Close } from 'ui/svg/icons/Close'
 import { Confidentiality } from 'ui/svg/icons/Confidentiality'
 import { ConfidentialityDeprecated } from 'ui/svg/icons/Confidentiality_deprecated'
+import { Confirmation } from 'ui/svg/icons/Confirmation'
 import { Digital } from 'ui/svg/icons/Digital'
 import { Duo } from 'ui/svg/icons/Duo'
 import { DuoBold } from 'ui/svg/icons/DuoBold'
@@ -109,9 +110,7 @@ import { HandicapMotor } from 'ui/svg/icons/HandicapMotor'
 import { HandicapVisual } from 'ui/svg/icons/HandicapVisual'
 import { HappyFace } from 'ui/svg/icons/HappyFace'
 import { HappyFaceStars } from 'ui/svg/icons/HappyFaceStars'
-import { Confirmation } from 'ui/svg/icons/IdentityCheck/Confirmation'
-import { IdCard } from 'ui/svg/icons/IdentityCheck/IdCard'
-import { Profil } from 'ui/svg/icons/IdentityCheck/Profil'
+import { IdCard } from 'ui/svg/icons/IdCard'
 import { Info } from 'ui/svg/icons/Info'
 import { InfoDeprecated } from 'ui/svg/icons/Info_deprecated'
 import { Invalidate } from 'ui/svg/icons/Invalidate'
@@ -138,6 +137,7 @@ import { OfferPhysical } from 'ui/svg/icons/OfferPhysical'
 import { OrderPrice } from 'ui/svg/icons/OrderPrice'
 import { PhoneFilled } from 'ui/svg/icons/PhoneFilled'
 import { PlainArrowPrevious } from 'ui/svg/icons/PlainArrowPrevious'
+import { Profil } from 'ui/svg/icons/Profil'
 import { Profile } from 'ui/svg/icons/Profile'
 import { ProfileDeletion } from 'ui/svg/icons/ProfileDeletion'
 import { RequestSent } from 'ui/svg/icons/RequestSent'
