+++ src/features/identityCheck/atoms/StepButton.tsx
@@ -19,7 +19,7 @@ export const StepButton = ({ step, state, onPress }: Props) => {
   return (
     <Button activeOpacity={ACTIVE_OPACITY} onPress={onPress} disabled={state === 'disabled'}>
       <IconContainer>
-        <Icon size={getSpacing(12)} />
+        <Icon size={getSpacing(10)} />
       </IconContainer>
       <Typo.ButtonText>{label}</Typo.ButtonText>
       <CompletionContainer>
@@ -43,7 +43,7 @@ const Button = styled.TouchableOpacity<{ disabled: boolean }>((props) => ({
   alignItems: 'center',
 }))
 
-const IconContainer = styled.View({ padding: getSpacing(2) })
+const IconContainer = styled.View({ padding: getSpacing(4) })
 
 const CompletionContainer = styled.View({
   flex: 1,
