--- src/ui/svg/icons/IdentityCheck/Confirmation.tsx
@@ -1,20 +1,0 @@
-import * as React from 'react'
-import Svg, { Path } from 'react-native-svg'
-
-import { IconInterface } from 'ui/svg/icons/types'
-import { ColorsEnum } from 'ui/theme'
-
-export function Confirmation({
-  size = 32,
-  color = ColorsEnum.BLACK,
-  testID,
-}: IconInterface): JSX.Element {
-  return (
-    <Svg width={size} height={size} testID={testID} viewBox="0 0 48 48" fill={color}>
-      <Path
-        opacity={0.5}
-        d="M30.6 23.25a7.25 7.25 0 110 14.5 7.25 7.25 0 010-14.5zm0 1.5a5.75 5.75 0 100 11.5 5.75 5.75 0 100-11.5zm3.396 2.62a.75.75 0 01.232.951l-.058.095-3.6 5.04a.75.75 0 01-1.135.1l-.076-.087-2.02-2.7a.75.75 0 011.129-.982l.073.084 1.405 1.878 3.004-4.205a.75.75 0 011.046-.174zM12.59 10.25l14.49.004c1.276.138 2.226 1.221 2.217 2.437l-.007.159v7.06a.75.75 0 01-1.493.102l-.007-.102.002-7.116c.037-.487-.29-.92-.69-1.025L27 11.75l-14.351-.002c-.5.039-.882.445-.901.977l.002.125-.002 17.884c-.036.5.308.939.74 1.031l.102.015h8.11a.75.75 0 01.102 1.493l-.102.007-8.169-.002c-1.305-.103-2.296-1.199-2.287-2.442l.006-.156.002-17.776c-.095-1.306.84-2.45 2.17-2.636l.168-.018zm12.13 8.91a.75.75 0 01.102 1.493l-.102.007h-9.9a.75.75 0 01-.102-1.493l.102-.007h9.9zm0-3.87a.75.75 0 01.102 1.493l-.102.007h-9.9a.75.75 0 01-.102-1.493l.102-.007h9.9z"
-      />
-    </Svg>
-  )
-}
