--- src/ui/svg/icons/IdentityCheck/IdCard.tsx
@@ -1,20 +1,0 @@
-import * as React from 'react'
-import Svg, { Path } from 'react-native-svg'
-
-import { IconInterface } from 'ui/svg/icons/types'
-import { ColorsEnum } from 'ui/theme'
-
-export function IdCard({
-  size = 32,
-  color = ColorsEnum.BLACK,
-  testID,
-}: IconInterface): JSX.Element {
-  return (
-    <Svg width={size} height={size} testID={testID} viewBox="0 0 48 48" fill={color}>
-      <Path
-        opacity={0.5}
-        d="M35.75 13a2.75 2.75 0 012.745 2.582l.005.168v18a2.75 2.75 0 01-2.75 2.75.75.75 0 01-.102-1.494L35.75 35a1.25 1.25 0 001.243-1.123L37 33.75v-18a1.25 1.25 0 00-1.122-1.244l-.128-.006h-23a1.25 1.25 0 00-1.244 1.122l-.006.128v18c0 .646.492 1.18 1.122 1.243l.128.007h17.378a.75.75 0 01.102 1.493l-.102.007H12.75a2.752 2.752 0 01-2.745-2.583L10 33.75v-18a2.75 2.75 0 012.582-2.745L12.75 13h23zm-17.185 4.48a3.12 3.12 0 013.119 3.12v1.312c0 1.366-.88 2.527-2.102 2.95l-.057.017a5.614 5.614 0 013.379 1.969 5.592 5.592 0 011.215 2.711.736.736 0 01-.017.311l-.006.018a.715.715 0 01-.077.174l-.017.026a.369.369 0 01-.04.055l-.014.017c-.379.444-.81.838-1.287 1.177a.75.75 0 01-.869-1.222 5.56 5.56 0 00.637-.528l.138-.14-.014-.065a4.09 4.09 0 00-.97-1.767 4.107 4.107 0 00-2.81-1.312l-.208-.006a4.044 4.044 0 00-.57.039l-.036.006.036-.006a3.06 3.06 0 00-.133.021l.097-.015c-.07.01-.14.022-.21.036l.113-.021c-.058.01-.116.021-.173.034l.06-.013a4.324 4.324 0 00-.19.043l.13-.03c-.067.014-.133.03-.199.048l.07-.018a3.375 3.375 0 00-.203.057l.133-.039a4.885 4.885 0 00-.183.054l.05-.015a3.583 3.583 0 00-.175.058l.125-.043c-.074.024-.148.05-.22.078l.095-.035c-.06.021-.12.044-.179.068l.084-.033a3.97 3.97 0 00-.213.088l.13-.055c-.057.023-.113.048-.169.073l.039-.018c-.053.024-.106.05-.158.076l.12-.058c-.064.03-.126.06-.187.092l.067-.034a3.989 3.989 0 00-.159.084l.092-.05a4.319 4.319 0 00-.188.106l.096-.056c-.05.028-.099.057-.147.086l.05-.03c-.051.03-.103.063-.154.096l.104-.066a5.553 5.553 0 00-.18.117l.076-.05a3.029 3.029 0 00-.165.114l.09-.064a4.202 4.202 0 00-.152.11l.062-.046a3.3 3.3 0 00-.162.124l.1-.079c-.05.038-.098.076-.145.116l.045-.037c-.05.04-.1.082-.148.125l.103-.088a5.06 5.06 0 00-.156.136l.053-.048c-.046.04-.09.082-.135.125l.082-.077a3.452 3.452 0 00-.138.133l.056-.056a3.731 3.731 0 00-.139.142l.083-.086a3.708 3.708 0 00-.137.144l.054-.058a5.486 5.486 0 00-.126.14l.072-.082a4.626 4.626 0 00-.129.15l.057-.067c-.041.047-.08.096-.12.145l.063-.078a4.828 4.828 0 00-.124.159l.062-.08a2.307 2.307 0 00-.095.126l.033-.046c-.035.048-.07.097-.104.147l.07-.1a4.998 4.998 0 00-.14.207l.07-.107c-.036.055-.072.11-.106.167l.036-.06-.09.15.054-.09a4.65 4.65 0 00-.104.18l.05-.09a4.247 4.247 0 00-.198.393l-.016.038.016-.038a4.25 4.25 0 00-.23.68c.17.175.348.336.534.485l.212.16a5.544 5.544 0 002.986 1.046l.275.007c.628 0 1.24-.103 1.817-.304a.75.75 0 01.491 1.418 7.025 7.025 0 01-1.969.378l-.34.008a7.076 7.076 0 01-2.739-.55 7.067 7.067 0 01-2.095-1.36 6.94 6.94 0 01-.55-.573l-.013-.016a.647.647 0 01-.037-.05l.05.066a.744.744 0 01-.08-.113l-.007-.013-.007-.013c-.011-.02-.02-.041-.03-.063l-.025-.07-.005-.02a.718.718 0 01-.018-.311 5.615 5.615 0 012.445-3.821 5.594 5.594 0 011.949-.819l.203-.037a3.12 3.12 0 01-2.16-2.969V20.6a3.12 3.12 0 013.12-3.12zm0 1.5a1.62 1.62 0 00-1.62 1.62v1.312a1.62 1.62 0 003.239 0V20.6a1.62 1.62 0 00-1.62-1.62zm15.757 2.944a.75.75 0 01.102 1.494l-.102.006h-7.345a.75.75 0 01-.101-1.493l.101-.007h7.345zm0-3.656a.75.75 0 01.102 1.493l-.102.007h-7.345a.75.75 0 01-.101-1.493l.101-.007h7.345z"
-      />
-    </Svg>
-  )
-}
