--- src/ui/svg/icons/IdentityCheck/Plus.tsx
@@ -1,13 +1,0 @@
-import * as React from 'react'
-import Svg, { Path } from 'react-native-svg'
-
-import { IconInterface } from 'ui/svg/icons/types'
-import { ColorsEnum } from 'ui/theme'
-
-export function Plus({ size = 32, color = ColorsEnum.BLACK, testID }: IconInterface): JSX.Element {
-  return (
-    <Svg width={size} height={size} testID={testID} viewBox="0 0 14 14" fill={color}>
-      <Path d="M7 14C10.866 14 14 10.866 14 7C14 3.13401 10.866 0 7 0C3.13401 0 0 3.13401 0 7C0 10.866 3.13401 14 7 14ZM7.00064 3.36685C7.36883 3.3672 7.66702 3.66597 7.66666 4.03416L7.66415 6.3304L9.95074 6.33335C10.3189 6.33377 10.6171 6.63259 10.6166 7.00078C10.6162 7.36897 10.3174 7.6671 9.94922 7.66668L7.66634 7.66368L7.67488 9.95608C7.67608 10.3243 7.37857 10.6237 7.01038 10.6249C6.6422 10.6261 6.34275 10.3286 6.34155 9.96041L6.33317 7.6671L4.04491 7.67582C3.67672 7.67707 3.37723 7.37961 3.37598 7.01142C3.37473 6.64323 3.6722 6.34374 4.04038 6.3425L6.33077 6.3336L6.33333 4.03286C6.33369 3.66467 6.63245 3.36649 7.00064 3.36685Z" />
-    </Svg>
-  )
-}
