--- src/ui/svg/icons/IdentityCheck/Profil.tsx
@@ -1,20 +1,0 @@
-import * as React from 'react'
-import Svg, { Path } from 'react-native-svg'
-
-import { IconInterface } from 'ui/svg/icons/types'
-import { ColorsEnum } from 'ui/theme'
-
-export function Profil({
-  size = 32,
-  color = ColorsEnum.BLACK,
-  testID,
-}: IconInterface): JSX.Element {
-  return (
-    <Svg width={size} height={size} testID={testID} viewBox="0 0 48 48" fill={color}>
-      <Path
-        opacity={0.5}
-        d="M24.619 25c5.317 0 9.805 3.89 10.611 9.096a.74.74 0 01.003.21l-.003.02a.64.64 0 01-.016.076l-.011.04-.022.058a.4.4 0 01-.02.042c-.007.016-.015.03-.023.045s-.017.028-.026.042a.765.765 0 01-.036.049l-.018.022a13.96 13.96 0 01-2.494 2.278.75.75 0 01-.87-1.222 12.35 12.35 0 001.938-1.71l.055-.062-.038-.19a9.19 9.19 0 00-.286-1.035l-.042-.12.042.12a9.01 9.01 0 00-.22-.581l-.012-.03.012.03a9.284 9.284 0 00-.425-.881l-.042-.076.042.076a8.44 8.44 0 00-.253-.433l-.082-.13.082.13c-.19-.306-.399-.6-.622-.88-.02-.027-.041-.052-.062-.078l.062.078-.147-.18.085.102a7.987 7.987 0 00-.186-.22l.101.119a9.633 9.633 0 00-.19-.22l.089.1-.133-.15.045.05a10.415 10.415 0 00-.176-.19l.131.14a13.862 13.862 0 00-.204-.218l.073.078a7.64 7.64 0 00-.177-.183l.104.105c-.058-.06-.117-.118-.176-.177l.072.072-.176-.172.104.1c-.065-.063-.13-.125-.197-.187l.093.087a9.184 9.184 0 00-.404-.36c-.021-.02-.043-.038-.065-.056l.065.055a9.145 9.145 0 00-.634-.49l-.052-.037.052.037a8.628 8.628 0 00-.197-.136l.145.099a7.37 7.37 0 00-.234-.158l.089.06a8.34 8.34 0 00-.21-.137l.12.077a7.117 7.117 0 00-.205-.13l.085.053c-.085-.053-.17-.105-.256-.156l.17.103a6.028 6.028 0 00-.219-.131l.049.028-.176-.1.127.072a6.454 6.454 0 00-.257-.143l.13.071a8.825 8.825 0 00-.493-.255l-.107-.051.107.051a9.404 9.404 0 00-.501-.223l-.077-.031.077.031-.17-.067.093.036a13.943 13.943 0 00-.261-.1l.169.064a9.278 9.278 0 00-.595-.206l-.049-.015.05.015a9.446 9.446 0 00-.748-.197l-.09-.02.09.02-.223-.046.132.027a7.454 7.454 0 00-.26-.05l.128.023a8.64 8.64 0 00-.58-.092l-.086-.011.086.01a9.418 9.418 0 00-.242-.027l.156.017c-.089-.01-.177-.02-.266-.028l.11.01a10.22 10.22 0 00-.659-.042l-.273-.004a9.203 9.203 0 00-4.855 1.38 9.286 9.286 0 00-1.987 1.655c-.056.061-.11.123-.164.185a9.246 9.246 0 00-1.23 1.838l-.115.234a9.287 9.287 0 00-.679 2.002l-.038.191a12.428 12.428 0 002.794 2.295 12.373 12.373 0 002.392 1.103c.776.26 1.582.442 2.408.542l.075.009c.343.039.69.064 1.038.074l.36.005a12.19 12.19 0 003.988-.666.75.75 0 01.49 1.418 13.674 13.674 0 01-4.459.748h-.018a13.716 13.716 0 01-2.305-.193 13.7 13.7 0 01-8.133-4.621l-.017-.02-.035-.047.052.067a.85.85 0 01-.08-.112l-.02-.036-.026-.057-.013-.035-.017-.057a.64.64 0 01-.016-.076l-.003-.02a.74.74 0 01.003-.21c.059-.379.137-.75.233-1.114l.076-.271a10.731 10.731 0 012.023-3.809l.144-.17a10.807 10.807 0 013.526-2.693l.249-.115a10.66 10.66 0 011.291-.478l.268-.076c.894-.242 1.833-.37 2.8-.37zm0-15a5.607 5.607 0 015.608 5.607v2.69a5.608 5.608 0 01-11.215 0v-2.69A5.607 5.607 0 0124.62 10zm0 1.5a4.107 4.107 0 00-4.107 4.107v2.69a4.107 4.107 0 108.215 0v-2.69A4.108 4.108 0 0024.62 11.5z"
-      />
-    </Svg>
-  )
-}
